Use request_cart fixture in check intergration tests
